+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 ## dev
 
-- Use virtual environment's Python to check for pip's availability during reinstalling process
+- Check whether pip module exists in shared lib before performing any actions, such as `reinstall-all`.
+- Drop `setuptools` and `wheel` from the shared libraries. This results in less time consumption when the libraries are
+automatically upgraded.
+- Allow running `pip` with `pipx run`
+- Support PEP 723 run requirements in `pipx run`.
+- Imply `--include-apps` when running `pipx inject --include-deps`
+- Add `--with-suffix` for `pipx inject` command
+- `pipx install`: emit a warning when `--force` and `--python` were passed at the same time
 - Drop support for Python 3.7
 - Make usage message in `pipx run` show `package_or_url`, so extra will be printed out as well
 - Add `--force-reinstall` to pip arguments when `--force` was passed